requirements.txt
sounddevice
numpy
librosa
scipy
matplotlib
python-kasa
tzdata